requirements.txt
fastapi
uvicorn
pydantic
pytest-html
sqlalchemy
uvicorn
coverage
python-multipart
python-jose
passlib
pydantic[email]
psycopg2-binary
alembic
pytz
pandas
httpx
